implementation of seperation widget of grade_view
